PortAudio: Allow bit-perfect audio output for macOS (#108)
When QMPlay2 isn't configured to resample all audio content to a
user-specific sample rate it sends the demuxed and decoded audio stream
"as is" to the selected audio device. When using the PortAudio module a
stream is opened that is set to the content sample rate, but no attempt
is made to configure the device appropriately.
This patch introduces an initial implementation of a bit-perfect audio
option in the PortAudio module for macOS. After opening the PA stream,
platform-native APIs are called to switch the selected audio device to
the content's sample rate or to the most appropriate (closest) supported
sample rate. The initial device settings are cached and restored when
playback is stopped.

barracuda156: Some fixes added for Qt4 compat.